+++ Public/DNS/Nameserver/func_Add-OPNameserver.ps1
@@ -40,6 +40,6 @@ function Add-OPNameServer {
         }
     }
     catch {
-        throw $_.Exception.Message
+        Write-Error $_.Exception.Message
     }
 }