+++ Public/DNS/Nameserver/func_Get-OPNameserver.ps1
@@ -28,7 +28,7 @@ function Get-OPNameServer {
         }
     }
     catch {
-        throw $_.Exception.Message
+        Write-Error $_.Exception.Message
     }
     return $return_object
 }