+++ Public/DNS/Nameserver/func_Get-OPNameserverGroup.ps1
@@ -17,7 +17,7 @@ function Get-OPNameserverGroup {
         }
     }
     catch {
-        throw $_.Exception.Message
+        Write-Error $_.Exception.Message
     }
     return $return_object
 }