+++ Public/DNS/Nameserver/func_Remove-OPNameserver.ps1
@@ -14,6 +14,6 @@ function Remove-OPNameserver {
     }
 
     catch {
-        throw $_.Exception.Message
+        Write-Error $_.Exception.Message
     }
 }