+++ Public/DNS/Nameserver/func_Update-OPNameserver.ps1
@@ -25,6 +25,6 @@ function Update-OPNameserver {
         }
     }
     catch {
-        throw $_.Exception.Message
+        Write-Error $_.Exception.Message
     }
 }