+++ Public/DNS/Zone/func_New-OPZone.ps1
@@ -27,6 +27,6 @@ function New-OPZone {
         }
     }
     catch {
-        throw $_.Exception.Message
+        Write-Error $_.Exception.Message
     }
 }