+++ Public/Domain/AuthCode/func_New-OPAuthCode.ps1
@@ -16,7 +16,7 @@ function New-OPAuthCode {
         }
     }
     catch {
-        throw $_.Exception.Message
+        Write-Error $_.Exception.Message
     }
     return $return_object
 }