+++ Public/SSL/func_Get-OPSSLOrders.ps1
@@ -51,10 +51,10 @@ function Get-OPSSLOrders {
         }
     }
     catch {
-        throw $_.Exception.Message
+        Write-Error $_.Exception.Message
     }
     if ($return_object.count -eq 0) {
-        throw "No SSL Orders found!"
+        Write-Error "No SSL Orders found!"
     }
     return $return_object
 }