+++ Public/SSL/func_Get-OPSSLProducts.ps1
@@ -19,7 +19,7 @@ function Get-OPSSLProducts {
         }
     }
     catch {
-        throw $_.Exception.Message
+        Write-Error $_.Exception.Message
     }
     return $return_object
 }